+++ package/types/DataTablePaginationProps.ts
@@ -17,6 +17,7 @@ export type DataTablePaginationProps = (
       paginationText?: never;
       paginationWrapBreakpoint?: never;
       getPaginationControlProps?: never;
+      getPaginationItemProps?: never;
       renderPagination?: never;
     }
   | {
@@ -98,6 +99,11 @@ export type DataTablePaginationProps = (
        * Useful for improving accessibility.
        */
       getPaginationControlProps?: (control: 'first' | 'last' | 'previous' | 'next') => Record<string, unknown>;
+
+      /**
+       * Function that returns props object for pagination item (page number button).
+       */
+      getPaginationItemProps?: (page: number) => Record<string, unknown>;
 
       /**
        * Optional render function to replace the entire pagination component.